Refactor de code
