+++ .gitignore
@@ -233,4 +233,5 @@ $RECYCLE.BIN/
 # Windows shortcuts
 *.lnk
 /.idea/*
-/src/main/java/games/project/statitisques/client64Bit/
+/src/main/java/games/project/statistiques/client64Bit/
+/src/main/PythonFileStats/app/client64Bit/
