+++ src/main/java/games/project/statistiques/views/ViewGamesStats.java
@@ -1,0 +1,4 @@
+package games.project.statistiques.views;
+
+public class ViewGamesStats {
+}
