+++ src/main/java/games/project/statistiques/views/ViewStatsGlobal.java
@@ -1,0 +1,4 @@
+package games.project.statistiques.views;
+
+public class ViewStatsGlobal {
+}
