+++ src/main/java/games/project/statistiques/views/ViewStatsTournament.java
@@ -1,0 +1,5 @@
+package games.project.statistiques.views;
+
+public class ViewStatsTournament {
+
+}
