--- src/main/java/games/project/statitisques/views/ViewGamesStats.java
@@ -1,4 +1,0 @@
-package games.project.statitisques.views;
-
-public class ViewGamesStats {
-}
