--- src/main/java/games/project/statitisques/views/ViewStatsPlayers.java
@@ -1,4 +1,0 @@
-package games.project.statitisques.views;
-
-public class ViewStatsPlayers {
-}
