--- src/main/java/games/project/statitisques/views/ViewStatsTournament.java
@@ -1,5 +1,0 @@
-package games.project.statitisques.views;
-
-public class ViewStatsTournament {
-
-}
