+++ src/main/java/module-info.java
@@ -25,9 +25,9 @@ module games.project.launcher {
     exports games.project;
     opens games.project.motron.controller to javafx.fxml;
 
-    exports games.project.statitisques;
-    opens games.project.statitisques.controller to javafx.fxml;
-    exports games.project.statitisques.views;
+    exports games.project.statistiques;
+    opens games.project.statistiques.controller to javafx.fxml;
+    exports games.project.statistiques.views;
 
     exports games.project.space_invader;
 }
